backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.2.1
pymongo==4.5.0
motor==3.3.1
pydantic>=2.12.5
email-validator>=2.3.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
python-jose>=3.5.0
requests>=2.32.5
python-multipart>=0.0.21
httpx>=0.27.0
